refactor(ui): TyConnectionBar shell + extract useSerialAutoSave
TyConnectionBar
- New shared component owns the conn-bar visual (left primary stripe,
  bg gradient, divider) and exposes icon / status / fields / actions
  slots. Both FirmwareFlashConnection and SerialDebugConnectionBar
  consume it.
- Padding is unified at p-2.5 sm:p-3 across both pages — the previous
  drift (p-3/3.5 vs p-2/2.5) caused a small layout jitter when the
  sidebar route changed between Flash and Serial debug.
- The plug icon block also moves to a shared slot pattern so serial
  debug no longer looks like a denser, icon-less variant.

useSerialAutoSave
- Lift the 130 lines of flush-watermark / interval / activated /
  deactivated bookkeeping out of SerialDebugPage.vue into a composable
  alongside the feature's other use*.ts files. The page is now 25
  lines of declarative layout.
